requirements.txt
flask==3.0.0
python-dotenv==1.0.0
openai==1.3.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2